# Notes-Server/swagger/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "url" : "http://localhost:3000/api/v1"
  } ],
  "info" : {
    "description" : "Add, edit, manager and share notes between users",
    "version" : "0.1-oas3",
    "title" : "Notes"
  },
...
  "paths" : {
    "/api/v1/users/register" : {
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Register a new user",
        "responses" : {
          "200" : {
            "description" : "New user successfully registered",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "name" : {
                    "type" : "string"
                  },
                  "email" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/users/login" : {
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Login with user credentials",
        "responses" : {
          "200" : {
            "description" : "Successful login",
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "token" : {
                      "type" : "string"
                    },
                    "userId" : {
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "401" : {
            "description" : "User not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Wrong password",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "email" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/users/detail" : {
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Get a particular user details",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "User details",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/users/otherusers" : {
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Get other user details",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Other user details",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find user/No other users found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "required" : true
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Response" : {
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "User" : {
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}